pyproject.toml
[project]
name = "nyc-civics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
